SW-3932 Extract Mapbox token fetching into hook
In preparation for adding a new React component that renders maps, move the logic
for fetching and refreshing Mapbox API tokens into a reusable hook.
